Add files via upload
a8PicoCart firmware updated with the Texas Atari Computer Owners group logo.  Thank you to Peter Dell for creating the updated rom.h file with the graphic and to Robin Edwards for information on how to build the new firmware.  This custom firmware also would not have been possible without Ryan Edwards working in tandem with me to debug and figure out how to successfully build the firmware.